Moving code to BaseModule class
