+++ openghg/modules/_metstore.py
@@ -16,24 +16,7 @@ class METStore(BaseModule):
     _uuid = "9fcabd0c-9b68-4ab4-a116-bc30a4472d67"
 
     def __init__(self):
-        from Acquire.ObjectStore import get_datetime_now
-        from collections import defaultdict
-
-        self._creation_datetime = get_datetime_now()
-        self._stored = False
-
-        # We want to created a nested dictionary
-        def nested_dict():
-            return defaultdict(nested_dict)
-
-        self._datasource_table = nested_dict()
-        # Keyed by name - allows retrieval of UUID from name
-        self._datasource_names = {}  
-        # Keyed by UUID - allows retrieval of name by UUID
-        self._datasource_uuids = {}
-        # Hashes of retrieved data to ensure we aren't overwriting / duplicating
-        # already stored data. Keyed by hash
-        self._hashes = {}
+        super().__init__()
 
     def to_data(self) -> Dict:
         """Return a JSON-serialisable dictionary of object
