+++ openghg/modules/_obs_surface.py
@@ -13,26 +13,7 @@ class ObsSurface(BaseModule):
     _uuid = "da0b8b44-6f85-4d3c-b6a3-3dde34f6dea1"
 
     def __init__(self):
-        from Acquire.ObjectStore import get_datetime_now
-        from collections import defaultdict
-
-        self._creation_datetime = get_datetime_now()
-        self._stored = False
-
-        # We want to created a nested dictionary
-        def nested_dict():
-            return defaultdict(nested_dict)
-
-        # Stores metadata about the Datasource, keyed by site
-        self._datasource_table = nested_dict()
-        # Keyed by name - allows retrieval of UUID from name
-        self._datasource_names = {}
-        # Keyed by UUID - allows retrieval of name by UUID
-        self._datasource_uuids = {}
-        # Hashes of previously uploaded files
-        self._file_hashes = {}
-        # Keyed by UUID
-        self._rank_data = defaultdict(dict)
+        super().__init__()
 
     def to_data(self) -> Dict:
         """Return a JSON-serialisable dictionary of object
